Hey so I I got somewhere.
I started the MapCreatorFromDat class but I got stuck on how to use the
expression classes so I can't continue for meow. I used the same
algorithm that you did in MapCreatorFromTxt to assign HighArea and
LowArea values, I just can't figure out the expression and operation
classes to get the next number to seek, so the class isn't working right
now because of NullPointers. Any insights are appreciated.

I also fixed your analyzer class and got the other analyzer classes
working with correct values. I tested this using data from terrain.txt.
The numbers aren't as rounded as they are in the example, but they're
the same.

Oh, and I reformatted some of your code so it's a little easier to read
:)

Thanks for setting everything up Corey! I got some stuff done but we
need to collaborate on how to get MapCreatorFromDat working correctly
since I'm a little lost.

~Greg