JAVADOC & UML!!!
I added Javadoc to every class, method, and fields. There are a lot of
changes here. The html files can be found in the doc folder. I also
re-formatted more code to make it look nice and tidy.

Corey gets all the credit for completing the UML diagram. He emailed it
to me for this final push. This can be found in the folder where the
terrain files are. It is called "la5uml.png".

This is the final version of the program and will be uploaded to
elearning.



Cheers,

Greg